A2Q2: fixed VTK indexing error
Voxel states were being passed in the order:
  [(x0, y0, z0), (x0, y0, z1), ..., (x0, y0, zn), ...]

but VTK was expecting the order:
  [(x0, y0, z0), (x1, y0, z0), ..., (xn, y0, z0), ...]

so the result was a jumble of noise. This commit fixes order of the
scalars passed into the VTK rectilinear grid.

There is a still an issue with projection 8 (idx 7) having NaNs and
resulting in a full array of zero states. This will be addressed next.